moved getDirectivesFromDocument to transform file to make it easy to move it apollo-utilities later
